+++ CHANGELOG.md
@@ -1,7 +1,35 @@
 #Change Log
 
+## [v10.9.2](https://github.com/auth0/lock/tree/v10.9.2) (2017-01-11)
+[Full Changelog](https://github.com/auth0/lock/compare/v10.9.1...v10.9.2)
+
+**Fixed**
+- Bring back support for get profile in default mode [\#794](https://github.com/auth0/lock/pull/794) ([glena](https://github.com/glena))
+- Don't emit error when registering for event 'signin ready' [\#784](https://github.com/auth0/lock/pull/784) ([theopak](https://github.com/theopak))
+
+
+## [v10.9.1](https://github.com/auth0/lock/tree/v10.9.1) (2017-01-10)
+[Full Changelog](https://github.com/auth0/lock/compare/v10.9.0...v10.9.1)
+
+**Fixed**
+- Fix to comply legacy behaviour [\#787](https://github.com/auth0/lock/pull/787) ([glena](https://github.com/glena))
+- For legacy flow, the scope should default to openid [\#783](https://github.com/auth0/lock/pull/783) ([glena](https://github.com/glena))
+
+
+## [v10.9.0](https://github.com/auth0/lock/tree/v10.9.0) (2017-01-09)
+[Full Changelog](https://github.com/auth0/lock/compare/v10.8.0-beta.5...v10.9.0)
+
+**Added**
+- Migrate to auth0.js v8 [\#744](https://github.com/auth0/lock/pull/744) ([glena](https://github.com/glena))
+
+**Changed**
+- Removed browserify as dependency and removed process usage [\#779](https://github.com/auth0/lock/pull/779) ([glena](https://github.com/glena))
+- Auth0js v8 - configuration validation + default scope [\#775](https://github.com/auth0/lock/pull/775) ([glena](https://github.com/glena))
+
+
 ## [v10.8.1](https://github.com/auth0/lock/tree/v10.8.1) (2017-01-03)
 [Full Changelog](https://github.com/auth0/lock/compare/v10.8.0...v10.8.1)
+
 **Closed issues**
 - username/password login doesn't work with custom domains on the appliance [\#772](https://github.com/auth0/lock/issues/772)
 
@@ -11,6 +39,7 @@
 
 ## [v10.8.0](https://github.com/auth0/lock/tree/v10.8.0) (2017-01-02)
 [Full Changelog](https://github.com/auth0/lock/compare/v10.7.3...v10.8.0)
+
 **Closed issues**
 - Request to add user-facing error message. [\#751](https://github.com/auth0/lock/issues/751)
 - Please throw an error for invalid events [\#748](https://github.com/auth0/lock/issues/748)